% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsTest.R
\name{runsTest}
\alias{runsTest}
\title{runsTest}
\usage{
runsTest(model)
}
\arguments{
\item{model}{A linear regression model specified using base R lm.}
}
\value{
Prints the number of expected runs, number of actual runs, and whether autocorrelation is likely.
}
\description{
Examines model residuals for the presence of autocorrelation.
}
\examples{
data(example)
linreg <- lm(H ~ P, data=example) 
runsTest(linreg)

}

